Update Table.cs
This is my first pull request, so I apologize if I am rude in any way.
I found a bug in table.cs, in lines 77 & 79.
That is, when calling the GetObject method of RowTable.cs, it was not being called as getObject in Table.cs.
So I changed 'g' to 'G'. Please check this.